dx12 dump refactor
Move all dump code from Dx12ReplayConsumerBase
to Dx12DumpResources.

In Dx12ReplayConsumerBase and Dx12ReplayConsumer,
if it needs to do dump resources tasks, it becomes

if (options_.enable_dump_resources)
{
    GFXRECON_ASSERT(dump_resources_);
    dump_resources_->...(...);
}

or

if(options_.enable_dump_resources)
{
   GFXRECON_ASSERT(dump_resources_);
   auto dump_command_sets = dump_resources_->GetCommandListsForDumpResources(...);
   for (auto& command_set : dump_command_sets)
   {
       command_set.list->...();
   }
}
