+++ framework/decode/dx12_object_info.h
@@ -91,6 +91,19 @@ T* GetExtraInfo(DxObjectInfo* info)
     return nullptr;
 }
 
+template <typename T>
+const T* GetExtraInfo(const DxObjectInfo* info)
+{
+    if ((info != nullptr) && (info->extra_info != nullptr) && (info->extra_info->extra_info_type == T::kType))
+    {
+        return static_cast<T*>(info->extra_info.get());
+    }
+
+    GFXRECON_LOG_FATAL("%s object does not have an associated info structure", T::kObjectType);
+
+    return nullptr;
+}
+
 struct MappedMemoryInfo
 {
     uint32_t count{ 0 };     ///< Number of times that the memory has been mapped.
